requirements.txt
apispec==5.1.1
alembic~=1.7.7
Flask==1.1.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.5.3
marshmallow~=3.17.0
Flask-Cors==3.0.10
flake8==4.0.1
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
PyYAML==6.0
apispec-webframeworks==0.5.2
flask-swagger-ui==4.11.1
itsdangerous==1.1.0
Werkzeug==2.0.3
Jinja2==3.0.0
python-dotenv==0.21.0
psycopg2-binary==2.9.3
requests==2.25.1
SQLAlchemy==1.4.27
